Moving aws-exports to general folder since it's used by a lot of files
