+++ apps/backend/src/aws-exports.ts
